+++ packages/react/src/components/UIShell/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.30.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.29.0...@carbon-labs/react-ui-shell@0.30.0) (2025-04-02)
+
+
+### Features
+
+* **UIShell:** add HeaderPopover ([#518](https://github.com/carbon-design-system/carbon-labs/issues/518)) ([1032d8a](https://github.com/carbon-design-system/carbon-labs/commit/1032d8a703637df97d3914b7fa82769b76d0446b))
+
+
+
+
+
+# [0.29.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.28.0...@carbon-labs/react-ui-shell@0.29.0) (2025-03-26)
+
+
+### Features
+
+* **uishell:** add header story with components update carbon/react ([#465](https://github.com/carbon-design-system/carbon-labs/issues/465)) ([a20ab0c](https://github.com/carbon-design-system/carbon-labs/commit/a20ab0c571bf739cbb906dfd08134c1966d06e0b))
+
+
+
+
+
 # [0.28.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.27.0...@carbon-labs/react-ui-shell@0.28.0) (2025-03-13)